ALL-2835 - Extension Ecosystem enabler - cleanup 2
