+++ src/service/extensions/index.ts
@@ -1,0 +1,3 @@
+export * from './tatumsdk.container'
+export * from './tatumsdk.extensions.dto'
+
