+++ src/service/index.ts
@@ -5,3 +5,4 @@ export * from './notification'
 export * from './rpc'
 export * from './tatum'
 export * from './walletProvider'
+export * from './extensions'
